Fix .gitignore and Write 'RoomSearch' Razor Page
Fix .gitignore to ignore 'appsettings.json' file. Create 'RoomSearch' page that allows a user to enter a start date and an end date and see which room types are available for that time period. Also, include link to 'Book Now' which will redirect to the booking page that will be created next.